feat: implement npmjs-deps-fetcher service
Provisions an npmjs-deps-fetcher service implementing the following API:
- GET /healthcheck (used for liveness/readiness probes)
- GET /package/{packageName}/{packageVersion} (returns the NPM
  dependency tree of an NPM package at a given version)

The service is composed of 3 core packages available in the /internal
directory:
- handler, which implements the HTTP handler for the GET package deps endpoint,
- npm, which provides a client, a resolver and data types to communicate
  with the NPM registry,
- semver, which provides helpers to deal with semantic versioning.

A complete local developer experience is also defined, with a Makefile
that provides handy commands to quickly spin up the service, either
directly as an OS process (i.e. `make run`) or as a containerized application
(i.e. `make docker-run`).

A suite of integration tests are available in the /test directory and
can be run with `make test-int`. Additionally, unit-tests to
independently test the different units that composed the sytem can be
run with `make test`.