chore: add integration tests
Adds integration tests by running the application as a subprocess of the
integration test run and faking the registry upstream server.

The application now uses Viper for handling configuration which
simplifies parsing of time.Duration, but also permits to use command
flags. This latter feature is useful when running the application from
the integration tests.

The registry upstream server uses the httptest package to create a
simple server that serves requests using json files stored in a testdata
folder.